patch: fix game engine crashing when evaluating king taken board
When evaluating a node that isn't depth 0 and the king has been taken,
the minimax didn't end the game there and continued to play after the
king had been taken. This caused an error in the chess.js engine to
access a board index of `-1` when looking at the location of the king on
the board.